Move Http\ControllerFactory to Controller\ControllerFactory.
